updated with doubles for money

+++ src/person/interfaces/Person.java
@@ -13,7 +13,7 @@ public interface Person {
 
 	public abstract void msgFinishedEvent(Role r); //The location manager will send this message as the persons role leaves the building
 	
-	public abstract void msgAddMoney(int money); //give money back to the person's wallet 
+	public abstract void msgAddMoney(double money); //give money back to the person's wallet 
 
 	public abstract void msgAddEvent(SimEvent e); //add an event to the person's scheduler
