+++ src/person/test/mock/PersonMock.java
@@ -59,7 +59,7 @@ public class PersonMock extends Mock implements Person {
 		
 	}
 	@Override
-	public void msgAddMoney(int money) { //add money back to the person's wallet
+	public void msgAddMoney(double money) { //add money back to the person's wallet
 		log.add(new LoggedEvent("Cash has been added to your personagent's wallet to the tune of "+money+" dollars"));
 		wallet.setOnHand(money);
 	}	
@@ -85,28 +85,28 @@ public class PersonMock extends Mock implements Person {
 	}
 	public class Wallet {
 		
-		private int onHand;
-		private int inBank;
-		private int balance; 
+		private double onHand;
+		private double inBank;
+		private double balance; 
 
-		Wallet(int oh, int ib){
+		Wallet(double oh, double ib){
 			this.onHand = oh;
 			this.inBank = ib;
 			this.balance = oh + ib;
 		}
-		public int getOnHand(){
+		public double getOnHand(){
 			return onHand;
 		}
-		public int getInBank(){
+		public double getInBank(){
 			return inBank;
 		}
-		public int getBalance(){
+		public double getBalance(){
 			return balance;
 		}
-		public void setOnHand(int newAmount){
+		public void setOnHand(double newAmount){
 			onHand += newAmount;
 		}
-		public void setInBank(int newAmount){
+		public void setInBank(double newAmount){
 			inBank += newAmount;
 		}
 	}
